feat:表达式编辑器支持变量为函数和表达式&editor导出上下文变量获取方法
