+++ packages/amis-editor-core/src/util.ts
@@ -11,6 +11,8 @@ import type {Schema} from 'amis';
 import type {SchemaObject} from 'amis';
 import {assign, cloneDeep} from 'lodash';
 import {getGlobalData} from 'amis-theme-editor-helper';
+import {isExpression, resolveVariableAndFilter} from 'amis-core';
+import type {VariableItem} from 'amis-ui';
 
 const {
   guid,
@@ -1132,3 +1134,90 @@ export function deleteThemeConfigData(data: any) {
 
   return schemaData;
 }
+
+/**
+ * 从amis数据域中取变量数据
+ * @param node
+ * @param manager
+ * @returns
+ */
+export async function resolveVariablesFromScope(node: any, manager: any) {
+  await manager?.getContextSchemas(node);
+  // 获取当前组件内相关变量，如表单、增删改查
+  const dataPropsAsOptions: VariableItem[] = updateComponentContext(
+    (await manager?.dataSchema?.getDataPropsAsOptions()) ?? []
+  );
+
+  const variables: VariableItem[] =
+    manager?.variableManager?.getVariableFormulaOptions() || [];
+
+  return [...dataPropsAsOptions, ...variables].filter(
+    (item: any) => item.children?.length
+  );
+}
+
+/**
+ * 整合 props & amis数据域 中的 variables
+ * @param that  为组件的实例 this
+ **/
+export async function getVariables(that: any) {
+  let variablesArr: any[] = [];
+
+  const {variables, requiredDataPropsVariables} = that.props;
+  if (!variables || requiredDataPropsVariables) {
+    // 从amis数据域中取变量数据
+    const {node, manager} = that.props.formProps || that.props;
+    let vars = await resolveVariablesFromScope(node, manager);
+    if (Array.isArray(vars)) {
+      if (!that.isUnmount) {
+        variablesArr = vars;
+      }
+    }
+  }
+  if (variables) {
+    if (Array.isArray(variables)) {
+      variablesArr = [...variables, ...variablesArr];
+    } else if (typeof variables === 'function') {
+      variablesArr = [...variables(that), ...variablesArr];
+    } else if (isExpression(variables)) {
+      variablesArr = [
+        ...resolveVariableAndFilter(
+          that.props.variables as any,
+          that.props.data,
+          '| raw'
+        ),
+        ...variablesArr
+      ];
+    }
+  }
+
+  // 如果存在应用语言类型，则进行翻译
+  if (that.appLocale && that.appCorpusData) {
+    return translateSchema(variablesArr, that.appCorpusData);
+  }
+
+  return variablesArr;
+}
+
+/**
+ * 更新组件上下文中label为带层级说明
+ * @param variables 变量列表
+ * @returns
+ */
+export const updateComponentContext = (variables: any[]) => {
+  const items = [...variables];
+  const idx = items.findIndex(item => item.label === '组件上下文');
+  if (~idx) {
+    items.splice(idx, 1, {
+      ...items[idx],
+      children: items[idx].children.map((child: any, index: number) => ({
+        ...child,
+        label:
+          index === 0
+            ? `当前数据域${child.label ? '(' + child.label + ')' : ''}`
+            : `上${index}层${child.label ? '(' + child.label + ')' : ''}`
+      }))
+    });
+  }
+  return items;
+};
