+++ packages/amis-editor/src/renderer/FormulaControl.tsx
@@ -3,16 +3,12 @@
  */
 
 import React from 'react';
-import debounce from 'lodash/debounce';
-import uniqBy from 'lodash/uniqBy';
 import isNumber from 'lodash/isNumber';
 import isBoolean from 'lodash/isBoolean';
 import isPlainObject from 'lodash/isPlainObject';
 import isArray from 'lodash/isArray';
 import isString from 'lodash/isString';
-import isEqual from 'lodash/isEqual';
 import omit from 'lodash/omit';
-import last from 'lodash/last';
 import cx from 'classnames';
 import {
   FormItem,
@@ -34,11 +30,11 @@ import type {
   VariableItem,
   FuncGroup
 } from 'amis-ui/lib/components/formula/Editor';
-import {dataMapping, FormControlProps} from 'amis-core';
+import {FormControlProps} from 'amis-core';
 import type {BaseEventContext} from 'amis-editor-core';
 import {EditorManager} from 'amis-editor-core';
 import {reaction} from 'mobx';
-import {getVariables} from './textarea-formula/utils';
+import {getVariables} from 'amis-editor-core';
 
 export enum FormulaDateType {
   NotDate, // 不是时间类
