+++ packages/amis-editor/src/util.ts
@@ -42,26 +42,3 @@ export const schemaToArray = (value: any) => {
 export const schemaArrayFormat = (value: any) => {
   return value && Array.isArray(value) && value.length === 1 ? value[0] : value;
 };
-
-/**
- * 更新组件上下文中label为带层级说明
- * @param variables 变量列表
- * @returns
- */
-export const updateComponentContext = (variables: any[]) => {
-  const items = [...variables];
-  const idx = items.findIndex(item => item.label === '组件上下文');
-  if (~idx) {
-    items.splice(idx, 1, {
-      ...items[idx],
-      children: items[idx].children.map((child: any, index: number) => ({
-        ...child,
-        label:
-          index === 0
-            ? `当前数据域${child.label ? '(' + child.label + ')' : ''}`
-            : `上${index}层${child.label ? '(' + child.label + ')' : ''}`
-      }))
-    });
-  }
-  return items;
-};
